update item deployment scripts
